Separate login form and container
